WIP figured out how scenes work. created a start scene and have it transition to main board scene
